refactor: new IUser interface location
