+++ src/modules/users/services/GetUserByIdService.ts
@@ -1,4 +1,4 @@
-import { IUser } from './interfaces/IUser';
+import { IUser } from '../interfaces/IUser';
 import { IUsersRepository } from '../repositories/interfaces/IUsersRepository';
 import { IGetUserByIdService } from './interfaces/IGetUserByIdService';
