+++ src/modules/users/services/UpdateUserService.ts
@@ -1,4 +1,4 @@
-import { IUser } from './interfaces/IUser';
+import { IUser } from '../interfaces/IUser';
 import { IUsersRepository } from '../repositories/interfaces/IUsersRepository';
 import { IUpdateUserService } from './interfaces/IUpdateUserService';
