+++ src/modules/users/services/interfaces/ICreateUserService.ts
@@ -1,4 +1,4 @@
-import { IUser } from './IUser';
+import { IUser } from '../../interfaces/IUser';
 
 export interface ICreateUserService {
   execute(user: IUser): Promise<void>;
