+++ src/modules/users/services/interfaces/IGetAllUsersService.ts
@@ -1,4 +1,4 @@
-import { IUser } from './IUser';
+import { IUser } from '../../interfaces/IUser';
 
 export interface IGetAllUsersService {
   execute(page: number): Promise<IUser[]>;
