+++ src/modules/users/services/interfaces/IGetUserByIdService.ts
@@ -1,4 +1,4 @@
-import { IUser } from './IUser';
+import { IUser } from '../../interfaces/IUser';
 
 export interface IGetUserByIdService {
   execute(id: string): Promise<IUser>;
