+++ src/modules/users/services/interfaces/IUpdateUserService.ts
@@ -1,4 +1,4 @@
-import { IUser } from './IUser';
+import { IUser } from '../../interfaces/IUser';
 
 export interface IUpdateUserService {
   execute(user: IUser): Promise<void>;
